+++ apps/forms/app/(dev)/playground/[slug]/page.tsx
@@ -1,0 +1,34 @@
+import { createServerComponentClient } from "@/lib/supabase/server";
+import { cookies } from "next/headers";
+import { notFound, redirect } from "next/navigation";
+import Playground from "@/scaffolds/playground";
+
+export default async function SharedPlaygroundPage({
+  params,
+}: {
+  params: {
+    slug: string;
+  };
+}) {
+  const { slug } = params;
+  const cookieStore = cookies();
+  const supabase = createServerComponentClient(cookieStore);
+
+  const { data: _gist } = await supabase
+    .from("playground_gist")
+    .select()
+    .eq("slug", slug)
+    .single();
+
+  if (!_gist) {
+    return redirect("/playground");
+  }
+
+  const { gist } = _gist;
+
+  return (
+    <main>
+      <Playground initial={JSON.stringify(gist, null, 2)} slug={slug} />
+    </main>
+  );
+}
