+++ apps/forms/app/(dev)/playground/share/route.ts
@@ -1,0 +1,22 @@
+import { createRouteHandlerClient } from "@/lib/supabase/server";
+import assert from "assert";
+import { cookies } from "next/headers";
+import { NextRequest, NextResponse } from "next/server";
+
+export async function POST(req: NextRequest) {
+  const body = await req.json();
+  const cookieStore = cookies();
+  const supabase = createRouteHandlerClient(cookieStore);
+
+  const { data } = await supabase
+    .from("playground_gist")
+    .insert({
+      gist: body,
+    })
+    .select()
+    .single();
+
+  assert(data, "Failed to create a new playground gist");
+
+  return NextResponse.json(data);
+}
