+++ apps/forms/scaffolds/playground/index.ts
@@ -1,0 +1,1 @@
+export { Playground as default } from "./playground";
